Modified MachineLearningGameValues to use chosenGameNightValues as a starting point for training

+++ src/Main/Models/Logic/IGameNightValues.java
@@ -6,11 +6,11 @@ package Main.Models.Logic;
 public interface IGameNightValues {
   double ownerHasPlayedGameLessThanFiveTimes();
 
-  int allPlayersHaveTriedGame();
+  double allPlayersHaveTriedGame();
 
   double playerRating(double rating);
 
-  int allPlayersHaveNotPlayedGameSinceTimeLimit();
+  double allPlayersHaveNotPlayedGameSinceTimeLimit();
 
   /**
    *
